+++ mysql_perf.py
@@ -228,9 +228,13 @@ def mysql_stat(server, args_array, **kwargs):
     json_fmt = args_array.get("-j", False)
     no_std = args_array.get("-z", False)
     mode = "w"
+    indent = 4
 
     if args_array.get("-a", False):
         mode = "a"
+
+    if args_array.get("-f", False):
+        indent = None
 
     # List of performance statistics to be checked.
     perf_list = ["indb_buf_data", "indb_buf_tot", "indb_buf_data_pct",
@@ -245,7 +249,8 @@ def mysql_stat(server, args_array, **kwargs):
     # Loop iteration based on the -n option.
     for x in range(0, int(args_array["-n"])):
         mysql_stat_run(server, perf_list, db_tbl=db_tbl, ofile=ofile,
-                       json_fmt=json_fmt, no_std=no_std, mode=mode, **kwargs)
+                       json_fmt=json_fmt, no_std=no_std, mode=mode,
+                       indent=indent, **kwargs)
 
         # Do not sleep on the last loop.
         if x != int(args_array["-n"]) - 1: