+++ mysql_perf.py
@@ -282,8 +282,8 @@ def run_program(args_array, func_dict, **kwargs):
         mongo = gen_libs.load_module(args_array["-m"], args_array["-d"])
 
     # Intersect args_array and func_dict to determine which functions to call.
-    for x in set(args_array.keys()) & set(func_dict.keys()):
-        func_dict[x](server, args_array, class_cfg=mongo, **kwargs)
+    for opt in set(args_array.keys()) & set(func_dict.keys()):
+        func_dict[opt](server, args_array, class_cfg=mongo, **kwargs)
 
     cmds_gen.disconnect([server])