+++ test/unit/mysql_perf/mysql_stat.py
@@ -29,7 +29,6 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_perf
-import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
@@ -67,6 +66,17 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
+        test_interval_negative -> Test with -b option set to negative.
+        test_interval_zero -> Test with -b option set to zero.
+        test_interval_one -> Test with -b option set to one.
+        test_interval_two -> Test with -b option set to > one.
+        test_loop_negative -> Test with -n option set to negative number.
+        test_zero_loop -> Test with -n option set to zero.
+        test_json_flat -> Test with flatten indentation for JSON.
+        test_json_indent -> Test with default indentation for JSON.
+        test_file_write -> Test with setting file write.
+        test_file_append -> Test with setting file append.
+        test_multi_loop -> Test with multiple loops.
         test_default -> Test with default settings.
 
     """
@@ -84,6 +94,159 @@ class UnitTest(unittest.TestCase):
         self.server = Server()
         self.args_array = {"-n": 1, "-b": 1}
         self.args_array2 = {"-n": 3, "-b": 1}
+        self.args_array3 = {"-n": 1, "-b": 1, "-a": True}
+        self.args_array4 = {"-n": 1, "-b": 1, "-f": True}
+        self.args_array5 = {"-n": 0, "-b": 1}
+        self.args_array6 = {"-n": 2, "-b": 2}
+        self.args_array7 = {"-n": 2, "-b": 1}
+        self.args_array8 = {"-n": 2, "-b": 0}
+        self.args_array9 = {"-n": 2, "-b": -1}
+        self.args_array10 = {"-n": -1, "-b": 1}
+
+    @unittest.skip("Error produced with negative -b option")
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_interval_negative(self, mock_process):
+
+        """Function:  test_interval_negative
+
+        Description:  Test with -b option set to negative.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array9))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_interval_zero(self, mock_process):
+
+        """Function:  test_interval_zero
+
+        Description:  Test with -b option set to zero.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array8))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_interval_one(self, mock_process):
+
+        """Function:  test_interval_two
+
+        Description:  Test with -b option set to one.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array7))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_interval_two(self, mock_process):
+
+        """Function:  test_interval_two
+
+        Description:  Test with -b option set to > one.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array6))
+
+    def test_loop_negative(self):
+
+        """Function:  test_loop_negative
+
+        Description:  Test with -n option set to negative number.
+
+        Arguments:
+
+        """
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array10))
+
+    def test_zero_loop(self):
+
+        """Function:  test_zero_loop
+
+        Description:  Test with -n option set to zero.
+
+        Arguments:
+
+        """
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array5))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_json_flat(self, mock_process):
+
+        """Function:  test_json_flat
+
+        Description:  Test with flatten indentation for JSON.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array4))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_json_indent(self, mock_process):
+
+        """Function:  test_json_indent
+
+        Description:  Test with default indentation for JSON.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_file_write(self, mock_process):
+
+        """Function:  test_file_write
+
+        Description:  Test with setting file write.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array3))
+
+    @mock.patch("mysql_perf.mysql_stat_run")
+    def test_file_append(self, mock_process):
+
+        """Function:  test_file_append
+
+        Description:  Test with setting file append.
+
+        Arguments:
+
+        """
+
+        mock_process.return_value = True
+
+        self.assertFalse(mysql_perf.mysql_stat(self.server, self.args_array3))
 
     @mock.patch("mysql_perf.mysql_stat_run")
     def test_multi_loop(self, mock_process):